middleware/requirements.txt
bidict==0.23.1
bitarray==2.9.2
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ffmpeg==1.4
Flask==3.0.3
Flask-SocketIO==5.3.6
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
netifaces==0.10.6
numpy==1.26.4
opencv-python==4.9.0.80
psutil==5.9.8
PyAudio==0.2.14
pycryptodome==3.20.0
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
setuptools==69.5.1
simple-websocket==1.0.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.2
wheel==0.43.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3
